app/requirements.txt
fastapi
uvicorn
langchain_together
langchain-huggingface
langchain_community
scikit-learn
einops
pandas
joblib
nomic
fpdf
python-docx
cloudinary